Move challenge generation to `plonk_proof.rs`
