+++ src/plonk_proof.rs
@@ -1,6 +1,7 @@
-use anyhow::Result;
+use anyhow::{anyhow, bail, ensure, Result};
 
-use crate::{AffinePoint, AffinePointTarget, Curve, Field, PartialWitness, Target};
+use crate::plonk_challenger::Challenger;
+use crate::{AffinePoint, AffinePointTarget, Curve, Field, PartialWitness, Target, SECURITY_BITS};
 
 #[derive(Debug, Clone, Copy)]
 pub struct SchnorrProof<C: Curve> {
@@ -55,6 +56,70 @@ impl<C: Curve> Proof<C> {
         ]
         .concat()
     }
+
+    // Computes all challenges used in the proof verification.
+    pub fn get_challenges(&self) -> Result<ProofChallenge<C>> {
+        let mut challenger = Challenger::new(SECURITY_BITS);
+        let error_msg = "Conversion from base to scalar field failed.";
+        challenger.observe_affine_points(&self.c_wires);
+        let (beta_bf, gamma_bf) = challenger.get_2_challenges();
+        let beta = C::try_convert_b2s(beta_bf).map_err(|_| anyhow!(error_msg))?;
+        let gamma = C::try_convert_b2s(gamma_bf).map_err(|_| anyhow!(error_msg))?;
+        challenger.observe_affine_point(self.c_plonk_z);
+        let alpha_bf = challenger.get_challenge();
+        let alpha = C::try_convert_b2s(alpha_bf).map_err(|_| anyhow!(error_msg))?;
+        challenger.observe_affine_points(&self.c_plonk_t);
+        let zeta_bf = challenger.get_challenge();
+        let zeta = C::try_convert_b2s(zeta_bf).map_err(|_| anyhow!(error_msg))?;
+        for os in self.all_opening_sets().iter() {
+            for &f in os.to_vec().iter() {
+                challenger.observe_element(C::try_convert_s2b(f).map_err(|_| anyhow!(error_msg))?);
+            }
+        }
+        let (v_bf, u_bf, u_scaling_bf) = challenger.get_3_challenges();
+        let v = C::try_convert_b2s(v_bf).map_err(|_| anyhow!(error_msg))?;
+        let u = C::try_convert_b2s(u_bf).map_err(|_| anyhow!(error_msg))?;
+        let u_scaling = C::try_convert_b2s(u_scaling_bf).map_err(|_| anyhow!(error_msg))?;
+
+        // Compute IPA challenges.
+        let mut ipa_challenges = Vec::new();
+        for i in 0..self.halo_l.len() {
+            challenger.observe_affine_points(&[self.halo_l[i], self.halo_r[i]]);
+            let l_challenge = challenger.get_challenge();
+            ipa_challenges.push(C::try_convert_b2s(l_challenge).map_err(|_| anyhow!(error_msg))?);
+        }
+
+        // Compute challenge for Schnorr protocol.
+        challenger.observe_affine_point(self.schnorr_proof.r);
+        let schnorr_challenge_bf = challenger.get_challenge();
+        let schnorr_challenge =
+            C::try_convert_b2s(schnorr_challenge_bf).map_err(|_| anyhow!(error_msg))?;
+
+        Ok(ProofChallenge {
+            beta,
+            gamma,
+            alpha,
+            zeta,
+            v,
+            u,
+            u_scaling,
+            ipa_challenges,
+            schnorr_challenge,
+        })
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ProofChallenge<C: Curve> {
+    pub beta: C::ScalarField,
+    pub gamma: C::ScalarField,
+    pub alpha: C::ScalarField,
+    pub zeta: C::ScalarField,
+    pub v: C::ScalarField,
+    pub u: C::ScalarField,
+    pub u_scaling: C::ScalarField,
+    pub ipa_challenges: Vec<C::ScalarField>,
+    pub schnorr_challenge: C::ScalarField,
 }
 
 pub struct ProofTarget {
